requirements.txt
argparse==1.2.1
beautifulsoup4==4.4.1
Django==1.9.12
django-allauth==0.28.0
django-analytical==1.0.0
django-autoslug==1.9.3
django-braces==1.8.1
django-cas-client==1.2.0
django-crispy-forms==1.5.2
django-filter==0.12.0
django-gravatar2==1.3.0
django-haystack==2.4.1
django-localflavor==1.2
django-model-utils==2.4
django-multiselectfield==0.1.3
django-randomslugfield==0.3.0
django-ratelimit==0.6.0
django-redis-cache==1.6.4
djangorestframework==3.3.2
flake8==2.5.0
gunicorn==19.6.0
Markdown==2.6.5
mccabe==0.3.1
MySQL-python==1.2.5
oauthlib==2.0.0
pep8==1.6.2
pyflakes==1.0.0
python-openid==2.2.5
redis==2.10.5
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
wsgiref==0.1.2
wheel==0.26.0
Whoosh==2.7.0
